+++ core/unittest/cms/common/impl/ExpiredMap_test.cpp
@@ -14,11 +14,11 @@ TEST(CommonExpiredMapTest, Normal) {
         return true;
     };
     ExpiredMap<int, int> data{microseconds{1000}};
-    EXPECT_EQ(data.Size(), 0);
+    EXPECT_EQ(data.Size(), size_t(0));
     EXPECT_TRUE(data.IsEmpty());
 
     EXPECT_EQ(data.Compute(1, fnCompute).v, 1);
-    EXPECT_EQ(1, data.Size());
+    EXPECT_EQ(size_t(1), data.Size());
     EXPECT_FALSE(data.IsEmpty());
 
     // 未过期，新值无效
@@ -27,7 +27,7 @@ TEST(CommonExpiredMapTest, Normal) {
         return true;
     };
     EXPECT_EQ(data.Compute(1, fnCompute1).v, 1);
-    EXPECT_EQ(1, data.Size());
+    EXPECT_EQ(size_t(1), data.Size());
     EXPECT_FALSE(data.IsEmpty());
 
     std::this_thread::sleep_for(microseconds{1100});
@@ -37,6 +37,6 @@ TEST(CommonExpiredMapTest, Normal) {
         return true;
     };
     EXPECT_EQ(data.Compute(1, fnCompute2).v, 2);
-    EXPECT_EQ(1, data.Size());
+    EXPECT_EQ(size_t(1), data.Size());
     EXPECT_FALSE(data.IsEmpty());
 }
